+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/snsattributes_test.go
@@ -1,0 +1,75 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package otelaws
+
+import (
+	"context"
+	"testing"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/sns"
+	"github.com/aws/smithy-go/middleware"
+	"github.com/stretchr/testify/assert"
+
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
+)
+
+func TestSNSPublish(t *testing.T) {
+	cases := map[string]struct {
+		input       middleware.InitializeInput
+		expectation string
+	}{
+		"When publish input with target arn is passed": {
+			middleware.InitializeInput{
+				Parameters: &sns.PublishInput{
+					TargetArn: aws.String("arn:aws:sns:us-east-1:0000000000:test-target-arn"),
+				},
+			},
+			"test-target-arn",
+		},
+		"When publish input with topic arn is passed": {
+			middleware.InitializeInput{
+				Parameters: &sns.PublishInput{
+					TopicArn: aws.String("arn:aws:sns:us-east-1:0000000000:test-topic-arn"),
+				},
+			},
+			"test-topic-arn",
+		},
+		"When publish input with a phone number is passed": {
+			middleware.InitializeInput{
+				Parameters: &sns.PublishInput{
+					PhoneNumber: aws.String("+4900000000000"),
+				},
+			},
+			"+4900000000000",
+		},
+		"When publish batch input is passed": {
+			middleware.InitializeInput{
+				Parameters: &sns.PublishBatchInput{
+					TopicArn: aws.String("arn:aws:sns:us-east-1:0000000000:test-topic-arn"),
+				},
+			},
+			"test-topic-arn",
+		},
+	}
+	for name, test := range cases {
+		t.Run(name, func(t *testing.T) {
+			attributes := SNSAttributeSetter(context.TODO(), test.input)
+
+			assert.Contains(t, attributes, semconv.MessagingDestinationName(test.expectation))
+			assert.Contains(t, attributes, semconv.MessagingDestinationKindTopic)
+		})
+	}
+}